[WEEX-224][android] WXDomObject Destory Method Should Be Called avoid Memory Leak
